refactor: the initial version of scheduler handler refactoring
